// frontend/src/Pages/landing-page/components/shared/types.ts
export interface TestimonialType {
  icon: JSX.Element
  icon_2: JSX.Element
  icon_3: JSX.Element
  icon_4: JSX.Element
  icon_5: JSX.Element
  icon_6: JSX.Element
  comment: string
  name: string
  location: string
}

export interface MenuType {
  sub_menu_1: string
  sub_menu_2: string
  sub_menu_3: string
  sub_menu_4: string
  title: string
}